requirements.txt
pymupdf4llm>=0.0.27
google-generativeai>=0.8.0
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0